Convert Deployments methods into global variables
These result in a nicer syntax for high-level pipelines (like serverlessPipeline).
